+++ extension/cart/transformToShopgateCart.js
@@ -227,7 +227,7 @@ function getCartItems (magentoCart, shopgateProducts) {
   if (magentoCart.coupon_code !== null && magentoCart.totals.hasOwnProperty('discount')) {
     const amount = Math.abs(parseFloat(magentoCart.totals.discount.value))
     const appliedDiscount = new AppliedDiscount(new SavedPrice(amount, 'fixed'))
-    appliedDiscount.code = magentoCart.coupon_code
+    appliedDiscount.code = magentoCart.coupon_code === '1' ? '' : magentoCart.coupon_code
     appliedDiscount.label = magentoCart.totals.discount.title
     const couponCartItem = new CartItem(magentoCart.coupon_code, 1, 'coupon', appliedDiscount)